// src/app/model/Invoice.ts
export class Invoice {
  issueDate: Date;
  total: number;
  paymentDay: Date;
  retention: number;
  difDays: number;
  Te: number;
  d: number;
  discount: number;
  initialCost: number;
  finalCost: number;
  netoValue: number;
  recibidoValue: number;
  entregadoValue: number;
  operationId: number;
  tcea: number;
}